внес часть правок по ревью

+++ profession/src/main/java/task/ProgrammerTask.java
@@ -3,7 +3,7 @@ package task;
 import lombok.NonNull;
 
 public interface ProgrammerTask {
-    Integer INITIAL_PROGRAMMER_TASK_LABOR_HOURS = 0;
+    Double INITIAL_PROGRAMMER_TASK_LABOR_HOURS = 0.0;
     String REDUCED_LABOR_HOURS_EXCEPTION = "Часы работы нельзя уменьшить";
 
     /**
@@ -17,7 +17,7 @@ public interface ProgrammerTask {
     /**
      * Задача сделана, переводится в статус "завершена", устанавливаются часы, но не меньше имеющихся по задаче.
      */
-    default void doneTask(@NonNull Task task, @NonNull Integer laborHours) throws LaborHoursException {
+    default void doneTask(@NonNull Task task, @NonNull Double laborHours) throws LaborHoursException {
         if (laborHours < task.getLaborHours()) {
             throw new LaborHoursException(REDUCED_LABOR_HOURS_EXCEPTION);
         }
